+++ warp10/src/main/java/io/warp10/script/ext/debug/DebugWarpScriptExtension.java
@@ -15,6 +15,7 @@ public class DebugWarpScriptExtension extends WarpScriptExtension {
     functions.put("STDOUT", new STDOUT("STDOUT"));
     functions.put("STDERR", new STDERR("STDERR"));
     functions.put("NOLOG", new NOLOG("NOLOG"));
+    functions.put("TDESCRIBE", new TDESCRIBE("TDESCRIBE"));
   }
   
   @Override
